⚒️ Forge: Refactor persistence logic out of God Object db.rs
- Extracted `PersistenceTracker` to `src/storage/index_persistence/tracker.rs`
- Extracted background persistence thread to `src/storage/index_persistence/background.rs`
- Updated `GallifreyDB::persist_indexes` to delegate to `storage::index_persistence::operations`
- Fixed type mismatch in background thread persistence calls
- Added documentation to new modules

This reduces the size and complexity of `src/db.rs` and improves separation of concerns.
